Add libre/foss flavor
This is the version without Google Play.

Because Google Play billing that makes income into the Citra Emulator Project works only with the original build of citra-emu,
and it doesn't work on devices that don't have Google Play,
and it requires a proprietary code.

It will always be the premium version because Google Play billing will not working. (I don't think it's good...)